+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Added `name` and `value` props to `Switch`, `CheckBox` and `RadioButton` components.
 -   Added `name` prop to `Autocomplete`, `DatePicker`, `TextField`, `Button`, `IconButton` and `Slider` components.
 -   Expose component default props in React `Component.defaultProps`
+-   A `Link` component without an `href` would result in a button looking like a link instead of an anchor.
+-   Added `typography` prop to `Link` component.
 
 ### Changed